+++ .github/workflows/deploy.yml
@@ -94,8 +94,7 @@ jobs:
       - name: Build and push
         uses: docker/build-push-action@v2
         with:
-          build-args:
-            API_URL: ${{API_URL}}
+          args: --build-arg API_URL=${{API_URL}}
           context: .
           push: true
           tags: ${{ steps.meta.outputs.tags }}
