--- CHANGELOG.md
@@ -1,7 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-### 0.1.2 (2021-12-15)
-
-### 0.1.1 (2021-12-02)
