+++ src/common/auth/service/Auth.ts
@@ -1,18 +1,16 @@
 import { AxiosInstance } from 'axios';
 import { AuthResponse } from './authResponse.interface';
 
-export const authByProvidersToken = async (axios: AxiosInstance, token: string, provider: string): Promise<AuthResponse> =>
+export const authByProvidersToken = async (axios: AxiosInstance, provider: string, token: string, state?: string): Promise<AuthResponse> =>
     axios.get<AuthResponse>(`/oauth/callback/${provider}`, {
         params: {
-            code: token
+            code: token,
+            state,
         }
     })
         .then(res => res.data);
 
 
 export const refreshToken = (axios: AxiosInstance, refresh_token: string): Promise<AuthResponse> =>
-    axios.post<AuthResponse>(`/oauth/access_token`, {
-        grant_type: "refresh_token",
-        refresh_token: refresh_token,
-    })
+    axios.post<AuthResponse>(`/api/refresh`, refresh_token)
         .then(res => res.data);
