+++ src/common/auth/service/authResponse.interface.ts
@@ -1,10 +1,10 @@
 
 export interface AuthResponse {
     username: string;
-    access_token: string;
-    refresh_token: string;
+    accessToken: string;
+    refreshToken: string;
 
     roles: string[];
-    expires_in: number;
-    token_type: string;
+    expiresIn: number;
+    tokenType: string;
 }
